added check

+++ GUI/connections.py
@@ -198,7 +198,7 @@ class MyConnections:
         currentIndex = self.dialog.configSelect.currentIndex()
         secPdu = self.mylib.GUIInterface_getSecuredPDU(currentIndex, byref(securedLen))
         print(securedLen)
-        if(securedLen != 0):
+        if(securedLen != c_byte(0)):
             self.mylib.GUIInterface_transmit(self.dialog.configSelect.currentIndex())
             self.dialog.tlog.info("Transmitted PDU ➡")
